requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
boto3==1.28.57
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1
